Split artifact phase out of executor.go
