+++ internal/job/executor.go
@@ -2272,97 +2272,6 @@ func (e *Executor) writeBatchScript(cmd string) (string, error) {
 
 }
 
-func (e *Executor) artifactPhase(ctx context.Context) error {
-	if e.AutomaticArtifactUploadPaths == "" {
-		return nil
-	}
-
-	spanName := e.implementationSpecificSpanName("artifacts", "artifact upload")
-	span, ctx := tracetools.StartSpanFromContext(ctx, spanName, e.ExecutorConfig.TracingBackend)
-	var err error
-	defer func() { span.FinishWithError(err) }()
-
-	err = e.preArtifactHooks(ctx)
-	if err != nil {
-		return err
-	}
-
-	err = e.uploadArtifacts(ctx)
-	if err != nil {
-		return err
-	}
-
-	err = e.postArtifactHooks(ctx)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// Run the pre-artifact hooks
-func (e *Executor) preArtifactHooks(ctx context.Context) error {
-	span, ctx := tracetools.StartSpanFromContext(ctx, "pre-artifact", e.ExecutorConfig.TracingBackend)
-	var err error
-	defer func() { span.FinishWithError(err) }()
-
-	if err = e.executeGlobalHook(ctx, "pre-artifact"); err != nil {
-		return err
-	}
-
-	if err = e.executeLocalHook(ctx, "pre-artifact"); err != nil {
-		return err
-	}
-
-	if err = e.executePluginHook(ctx, "pre-artifact", e.pluginCheckouts); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// Run the artifact upload command
-func (e *Executor) uploadArtifacts(ctx context.Context) error {
-	span, _ := tracetools.StartSpanFromContext(ctx, "artifact-upload", e.ExecutorConfig.TracingBackend)
-	var err error
-	defer func() { span.FinishWithError(err) }()
-
-	e.shell.Headerf("Uploading artifacts")
-	args := []string{"artifact", "upload", e.AutomaticArtifactUploadPaths}
-
-	// If blank, the upload destination is buildkite
-	if e.ArtifactUploadDestination != "" {
-		args = append(args, e.ArtifactUploadDestination)
-	}
-
-	if err = e.shell.Run(ctx, "buildkite-agent", args...); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// Run the post-artifact hooks
-func (e *Executor) postArtifactHooks(ctx context.Context) error {
-	span, _ := tracetools.StartSpanFromContext(ctx, "post-artifact", e.ExecutorConfig.TracingBackend)
-	var err error
-	defer func() { span.FinishWithError(err) }()
-
-	if err = e.executeGlobalHook(ctx, "post-artifact"); err != nil {
-		return err
-	}
-
-	if err = e.executeLocalHook(ctx, "post-artifact"); err != nil {
-		return err
-	}
-
-	if err = e.executePluginHook(ctx, "post-artifact", e.pluginCheckouts); err != nil {
-		return err
-	}
-
-	return nil
-}
-
 // Check for ignored env variables from the job runner. Some
 // env (for example, BUILDKITE_BUILD_PATH) can only be set from config or by hooks.
 // If these env are set at a pipeline level, we rewrite them to BUILDKITE_X_BUILD_PATH
